Remove commented lines of code * Change card int type raduis params to float type

+++ ui/decredmaterial/card.go
@@ -16,10 +16,10 @@ type Card struct {
 }
 
 type CornerRadius struct {
-	NE int
-	NW int
-	SE int
-	SW int
+	NE float32
+	NW float32
+	SE float32
+	SW float32
 }
 
 const (
@@ -44,10 +44,10 @@ func (c Card) Layout(gtx layout.Context, w layout.Widget) layout.Dimensions {
 			return c.Inset.Layout(gtx, func(gtx C) D {
 				return layout.Stack{}.Layout(gtx,
 					layout.Expanded(func(gtx C) D {
-						tr := float32(gtx.Px(unit.Dp(float32(c.Radius.NE))))
-						tl := float32(gtx.Px(unit.Dp(float32(c.Radius.NW))))
-						br := float32(gtx.Px(unit.Dp(float32(c.Radius.SE))))
-						bl := float32(gtx.Px(unit.Dp(float32(c.Radius.SW))))
+						tr := float32(gtx.Px(unit.Dp(c.Radius.NE)))
+						tl := float32(gtx.Px(unit.Dp(c.Radius.NW)))
+						br := float32(gtx.Px(unit.Dp(c.Radius.SE)))
+						bl := float32(gtx.Px(unit.Dp(c.Radius.SW)))
 						clip.RRect{
 							Rect: f32.Rectangle{Max: f32.Point{
 								X: float32(gtx.Constraints.Min.X),
